log: update logging on the ESP8266
Update the system tick pintout for the ESP8266.

+++ source/log.c
@@ -26,7 +26,7 @@ static void vfprint_dbg(const char *prefix, const char *fmt, va_list va)
 #ifdef AVR
 	fprintf(dbg_file, "[%lu]", (unsigned long)tick);
 #elif defined(ESP8266)
-	fprintf(dbg_file, "[%lli]", (long long)tick);
+	fprintf(dbg_file, "[%li]", (long)tick);
 #else
 	fprintf(dbg_file, "[%llu]", (unsigned long long)tick);
 #endif
